+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 * `[torch]` Toggle torch colours for certification purposes (redish, blueish, off).
+* `[calendar]` database structure
 
 ### Changed
 
@@ -79,7 +80,6 @@
 * `[calendar]` Added day window with list of events.
 * `[calendar]` main screen UI
 * `[calendar]` main screen navigation
-* `[calendar]` database structure
 * `[audio]` added incoming call ringtone
 * `[phonebook]` general support for groups in DB
 * `[phonebook]` proper displaying of favorites, ice, blocked flags